backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# LangChain & LangGraph - Industry-Standard Orchestration
langchain #==0.3.0
langchain-community #==0.3.0
langchain-core #==0.3.0
langgraph #==0.2.0
langchain-groq #==0.2.0

# Vector Store - Production-Quality Free Tier
qdrant-client==1.11.0

# Local Embeddings - Zero API Cost
sentence-transformers==3.0.0

# Web Search - AI-Optimized Search
tavily-python==0.5.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3

# Utilities
python-dotenv==1.0.0
pydantic==2.9.0
pydantic-settings==2.5.0

tenacity==8.2.3  # For retry decorators